TypeScript Configuration given an addition; Profile
